+++ CHANGELOG.md
@@ -5,8 +5,13 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+## [0.3.1] - 2019-04-18
 ### Added
 - ForceFlag: flag to support --force and -f
+
+### Changed
+- add features simplelog in docs.rs (hopefully)
 
 ## [0.3.0] - 2019-04-10
 ### Added